% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinedBoundaries.R
\name{ldNominalSig}
\alias{ldNominalSig}
\title{Calculate Nominal Significance, Error Spending Approach}
\usage{
ldNominalSig(alpha, tVec, cvec, locPeak, type = 2, initIntvl = c(1, 4))
}
\arguments{
\item{alpha}{original significance level.}

\item{tVec}{information vector.}

\item{cvec}{primary group sequential boundary.}

\item{locPeak}{location of maximum, a number between 1 and the number of interims.}

\item{type}{O'Brien- Fleming (OBF) type error spending funciton is 1, Pocock (POC) type error spending funciton is 2.}

\item{initIntvl}{computing paramter, a pair of numbers containing the end-points of the interval to be searched for the root.}
}
\value{
nominal significance of the secondary group sequential boundary.
}
\description{
Nominal significance for the secondary endpoint are calculated by using the error spending approach.
}
\details{
This function calculates the nominal significance level of any Lan-DeMets error spending boundary.
The original significance level is used to choose the initial searching range of the nominal significance.
}
\examples{
\dontrun{
require(mvtnorm)
require(ldbounds)
K <- 6;
tVec <- c(140,328,453,578,659,1080)/1080;
alpha <- 0.025;
cvec.obf <- ldbounds::ldBounds(t=tVec,iuse=c(1),alpha=c(alpha),sides = 1);
cvec <- cvec.obf$upper.bounds;
alphaprime <- ldNominalSig(alpha,tVec,cvec,locPeak=4,type=2,
      initIntvl=c(1,4))
}

}
\references{
Lan, K. K. G., and Demets, D. L. (1983). Discrete sequential boundaries for clinical trials. \emph{Biometrika} \bold{70}, 659-663.

Tamhane, A. C., Gou, J., Jennison, C., Mehta, C. R., and Curto, T. (2018). A gatekeeping procedure to test a primary and a secondary endpoint in a group sequential design with multiple interim looks. \emph{Biometrics}, 74, 40-48.
}
\seealso{
\code{nominalSig}, \code{secondaryBoundaryVecLD}
}
\author{
Jiangtao Gou
}
