% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinedBoundaries.R
\name{primaryBoundary}
\alias{primaryBoundary}
\title{Calculate Primary Boundaries, Standard Approach}
\usage{
primaryBoundary(gammaVec, alpha, type = 1, initIntvl = c(1, 4))
}
\arguments{
\item{gammaVec}{a vector of square root of information.}

\item{alpha}{significance level}

\item{type}{type of sequential procedure. OBF is 1, POC is 2.}

\item{initIntvl}{paramter for function uniroot (two numbers)}
}
\value{
original OBF and POC boundaries (primary endpoints) (a number, c_(K)).
}
\description{
Primary boundaries calculation of standard (original) OBF and POC.
}
\references{
Tamhane, A. C., Gou, J., Jennison, C., Mehta, C. R., and Curto, T. (2017+). A gatekeeping procedure to test a primary and a secondary endpoint in a group sequential design with multiple interim looks. \emph{Biometrics}, to appear.
}
\seealso{
\code{ldPrimaryBoundary}
}
\author{
Jiangtao Gou
}
