% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinedBoundaries.R
\name{secControl}
\alias{secControl}
\title{Difference between the Error Rate and Significance Level, Standard Approach}
\usage{
secControl(d, alpha, cvec, gammaVec, ExtrmLoc, type = 2)
}
\arguments{
\item{d}{boundary of secondary endpoint at the final look (a number, d_(K))}

\item{alpha}{targeted significance level for the secondary endpoint}

\item{cvec}{a vector of calculated primary boundaries}

\item{gammaVec}{square root of information}

\item{ExtrmLoc}{an integer between 1 and K, locate the maximum of type I error of secondary endpoint}

\item{type}{type of sequential procedures. Type 1 OBF d, Type 2 POC d.}
}
\value{
difference between alpha and the calculated error rate.
}
\description{
Calculate the difference between the error rate and significance level for the secondary endpoint, standard (original) approach.
}
\references{
Tamhane, A. C., Gou, J., Jennison, C., Mehta, C. R., and Curto, T. (2018). A gatekeeping procedure to test a primary and a secondary endpoint in a group sequential design with multiple interim looks. \emph{Biometrics}, 74(1), 40-48.
}
\seealso{
\code{ldSecControl}
}
\author{
Jiangtao Gou
}
